+++ volumes.tf
@@ -16,7 +16,7 @@
 
 # define the block volume
 resource "oci_core_volume" "CreateVolume" {
-  count               = var.host_volume_enabled == true ? var.tvd_participants : 0
+  count               = var.host_volume_enabled == true ? var.numberOf_labs : 0
   availability_domain = local.availability_domain
   compartment_id      = var.compartment_ocid
   display_name        = var.label_prefix == "none" ? format("${local.resource_shortname}-${var.host_name}%02d-volume", count.index) : format("${var.label_prefix}-${local.resource_shortname}-${var.host_name}%02d-volume", count.index)
@@ -34,7 +34,7 @@ resource "oci_core_volume" "CreateVolume" {
 
 # Create a volume attachement
 resource "oci_core_volume_attachment" "CreateVolumeAttachment" {
-  count           = var.host_volume_enabled == true ? var.tvd_participants : 0
+  count           = var.host_volume_enabled == true ? var.numberOf_labs : 0
   attachment_type = var.host_volume_attachment_type
   instance_id     = oci_core_instance.compute[count.index].id
   volume_id       = oci_core_volume.CreateVolume[count.index].id